create arrow_util.hpp/cpp instead of letting this live in arrow_wrapper.hpp
